requirements.txt
chainlit
groq
python-dotenv
pyodbc
